arbor-core/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
nats-py[nkeys]==2.7.2
redis[hiredis]==5.1.1
asyncpg==0.29.0
httpx==0.27.2
pydantic==2.9.2
aiofiles==24.1.0
python-multipart==0.0.9
structlog==24.4.0
zeroconf>=0.132.0
